+++ packages/camera/camera_android/android/src/main/java/io/flutter/plugins/camera/Camera.java
@@ -15,6 +15,7 @@ import android.hardware.camera2.CameraCaptureSession;
 import android.hardware.camera2.CameraDevice;
 import android.hardware.camera2.CameraManager;
 import android.hardware.camera2.CameraMetadata;
+import android.hardware.camera2.CaptureFailure;
 import android.hardware.camera2.CaptureRequest;
 import android.hardware.camera2.TotalCaptureResult;
 import android.hardware.camera2.params.OutputConfiguration;
@@ -682,11 +683,37 @@ class Camera
 
       // Trigger one capture to start AE sequence.
       captureSession.capture(
-          previewRequestBuilder.build(), cameraCaptureCallback, backgroundHandler);
+          previewRequestBuilder.build(),
+          createTriggerResetCallback(
+              CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
+              CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_IDLE),
+          backgroundHandler);
 
     } catch (CameraAccessException e) {
       e.printStackTrace();
     }
+  }
+
+  @VisibleForTesting
+  CameraCaptureSession.CaptureCallback createTriggerResetCallback(
+      final CaptureRequest.Key<Integer> triggerKey, final int triggerIdleValue) {
+    return new CameraCaptureSession.CaptureCallback() {
+      @Override
+      public void onCaptureCompleted(
+          @NonNull CameraCaptureSession session,
+          @NonNull CaptureRequest request,
+          @NonNull TotalCaptureResult result) {
+        previewRequestBuilder.set(triggerKey, triggerIdleValue);
+      }
+
+      @Override
+      public void onCaptureFailed(
+          @NonNull CameraCaptureSession session,
+          @NonNull CaptureRequest request,
+          @NonNull CaptureFailure failure) {
+        previewRequestBuilder.set(triggerKey, triggerIdleValue);
+      }
+    };
   }
 
   /**
@@ -795,7 +822,11 @@ class Camera
         CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_START);
 
     try {
-      captureSession.capture(previewRequestBuilder.build(), null, backgroundHandler);
+      captureSession.capture(
+          previewRequestBuilder.build(),
+          createTriggerResetCallback(
+              CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_IDLE),
+          backgroundHandler);
     } catch (CameraAccessException e) {
       String message =
           (e.getMessage() == null)