+++ packages/url_launcher/url_launcher/example/lib/encoding.dart
@@ -21,6 +21,7 @@ String? encodeQueryParameters(Map<String, String> params) {
       )
       .join('&');
 }
+
 // #enddocregion encode-query-parameters
 
 void main() => runApp(