+++ packages/camera/camera/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.12.0+2
 
 * Adds `videoOutputPath` support to `startVideoRecording`.
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 0.12.0+1